Impl IntoIterator for NonEmpty
Added implementations of `IntoIterator` for `NonEmpty`, `&NonEmpty`, and `&mut NonEmpty`.

While these implementations are not necessary to call `.iter` (as that previously worked via `Deref`), this does not work for `iter_mut` or `into_iter`. Additionally, this failed for any of the previous forms when using for loops, or trying to pass a `NonEmpty` into a fn that expects an `IntoIterator`.